+++ example/spec/parallel/lzardsBackup/LzardsBackupSpec.js
@@ -260,6 +260,8 @@ describe('The Lzards Backup Task ', () => {
       else {
         const lzardsGetPayload = new TextEncoder().encode(JSON.stringify({
           searchParams: {
+            provider: 'BOGUS_PROVIDER',
+            requestSummary: false,
             'metadata[collection]': `${collection.name}___${collection.version}`,
             'metadata[granuleId]': granuleId,
           },
@@ -285,6 +287,8 @@ describe('The Lzards Backup Task ', () => {
       else {
         const lzardsGetPayload = new TextEncoder().encode(JSON.stringify({
           searchParams: {
+            provider: 'BOGUS_PROVIDER',
+            requestSummary: false,
             pageLimit: 25,
             'metadata[provider]': provider,
             'metadata[createdAt][gte]': thirtyMinutesAgo,
@@ -312,6 +316,8 @@ describe('The Lzards Backup Task ', () => {
       else {
         const lzardsGetPayload = new TextEncoder().encode(JSON.stringify({
           searchParams: {
+            provider: 'BOGUS_PROVIDER',
+            requestSummary: false,
             'metadata[collection]': 'notBackedUpCollectionName',
             'metadata[granuleId]': granuleId,
           },